yolo-detection/requirements.txt
# AWS SDK
boto3==1.34.34
sagemaker==2.210.0

# YOLO and Computer Vision
ultralytics==8.1.0
opencv-python==4.9.0.80
torch==2.2.0
torchvision==0.17.0

# Image Processing
Pillow==10.2.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
